% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_mut_type}
\alias{build_mut_type}
\title{\strong{Build the final mutation type}}
\usage{
build_mut_type(mutgr, reference_genome, context_length, rev)
}
\arguments{
\item{mutgr}{GRanges object containing a set of single nucleotide variants}

\item{reference_genome}{Reference genome}

\item{context_length}{Parameter specified by the user to indicate the overall length of the mutation type - it is
used to compute how many nucleotides upstream and downstream the SNV base include in the mutation type}

\item{rev}{Parameter to indicate whether the reverse complement of the sequence containing the SNV must be computed
(default is FALSE)}
}
\value{
Vector with the mutation types `UP[REF>ALT]DOWN` of all the variants in the GRanges object with C or T as
mutated reference base.
}
\description{
\code{build_mut_type} function takes a GRanges object containing a set of single nucleotide variants retrieved from a VCF file,
the corresponding reference genome, the parameter \emph{context_length} specified by the user and the indication whether
the reverse complement of the sequence containing the SNV must be computed.\cr
This function returns a vector with the mutation types `UP[REF>ALT]DOWN` of all the variants in the GRanges
object, such that all mutation types have C or T as mutated reference base.\cr
The overall length of the mutation type is determined by the context_length parameter.\cr
This function is not exported to be used by the user but it is defined only to be exploited in the \link{mutation_type}
function.
}
\author{
Paola Maragno\cr Politecnico di Milano\cr Maintainer: Paola Maragno\cr E-Mail: <paola.maragno@mail.polimi.it>
}
