% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_table}
\alias{count_table}
\title{\strong{Count table of mutation types}}
\usage{
count_table(mut_types)
}
\arguments{
\item{mut_types}{Vector of mutation types `UP[REF>ALT]DOWN` returned by \link{mutation_type} function}
}
\value{
Count table reporting the number of mutations per mutation type.
}
\description{
\code{count_table} function summarizes the mutation types for the set of mutations into a count table
reporting the number of mutations per mutation type.
}
\examples{

mut_types <- c("GT[C>A]CA", "GG[C>G]TC", "CC[T>C]TC", "GT[C>T]GT", "TA[C>A]CG", "GT[C>A]CA",
"CA[C>G]CT", "CC[T>C]CT", "CA[T>C]AT", "TT[C>G]TC", "CC[T>C]CT", "CT[T>C]TG", "CC[T>C]CT", "TA[C>A]CG")
count_table(mut_types)

}
\author{
Paola Maragno\cr Politecnico di Milano\cr Maintainer: Paola Maragno\cr E-Mail: <paola.maragno@mail.polimi.it>
}
