% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graphical_summary}
\alias{graphical_summary}
\title{\strong{Graphical visualization of the mutation types with a frequancy higher than a threshold}}
\usage{
graphical_summary(c_table, freq, file_name)
}
\arguments{
\item{c_table}{Count table returned by \link{count_table} function that summarizes the mutation types contained in a VCF file}

\item{freq}{Threshold of frequency that a mutation type must have at least in order to be visualized in the barplot}

\item{file_name}{Name of the pdf file in which the function will plot its graphical output}
}
\value{
The function returns the name of the pdf file - that is named as chosen by the user and it is stored in the working
directory - showing the barplot visualization of the frequencies of the mutation types with a frequency at least equal to
\emph{freq} threshold.
}
\description{
\code{graphical_summary} function generates a pdf file showing the barplot visualization of all the mutation
types with a frequency higher than a threshold specified by the user.
}
\examples{

c_table <- data.frame(mut_types = c("AA[C>A]AA", "AA[C>A]AC", "AA[C>A]AG", "AA[C>A]AT", "AA[C>A]CA", "AA[C>A]CC",
 "AA[C>A]CT", "AA[C>A]GA", "AA[C>A]GC", "AA[C>A]GG", "AA[C>A]GT", "AA[C>A]TA", "AA[C>A]TC"),
 Freq = c(15,  13,  32,  44,  25,  5,  67,  21,  14,  42,  52,  21,  19))
graphical_summary(c_table, 30, "Mut_types_visualization_30.pdf")

}
\author{
Paola Maragno\cr Politecnico di Milano\cr Maintainer: Paola Maragno\cr E-Mail: <paola.maragno@mail.polimi.it>
}
