% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutation_type}
\alias{mutation_type}
\title{\strong{Determine the mutation type of a set of single nucleotide variants in a genome}}
\usage{
mutation_type(vcf_file, reference_genome, context_length)
}
\arguments{
\item{vcf_file}{VCF file containing a set of single nucleotide variants (like `chr22.vcf`)}

\item{reference_genome}{Reference genome}

\item{context_length}{Parameter specified by the user to indicate the overall length of the mutation type - it is
used to compute how many nucleotides upstream and downstream the SNV base include in the mutation type}
}
\value{
List of two elements: the mutation types `UP[REF>ALT]DOWN` vector with C or T as mutated reference base
and the GRanges object containing first the SNVs with C or T as mutated reference base and then the SNVs with G or A as
mutated reference base. In this way the position of each mutation type in the vector is the same of the corresponding SNV in
the GRanges object returned by the function.
}
\description{
\code{mutation_type} function takes a set of single nucleotide variants in VCF format, the corresponding reference
genome and a parameter \emph{context_length} specified by the user - that must be odd - and determines for each mutation
the corresponding mutation type `UP[REF>ALT]DOWN` such that all mutation types have C or T as mutated reference base.\cr
The overall length of the mutation type is determined by the context_length parameter.
}
\examples{

if (!'BSgenome.Hsapiens.UCSC.hg19' \%in\% installed.packages()) {
if (!requireNamespace("BiocManager", quietly = TRUE)) {
install.packages("BiocManager")
}
BiocManager::install('BSgenome.Hsapiens.UCSC.hg19')
}

library('BSgenome.Hsapiens.UCSC.hg19')
vcf_file <- system.file("extdata", "chr22.vcf.gz", package = "VariantAnnotation")
results <- mutation_type(vcf_file, BSgenome.Hsapiens.UCSC.hg19, 3)
mut_types <- results[[1]]
mutgr_reordered <- results[[2]]
mut_types[1:10]
mutgr_reordered[1:10]

}
\author{
Paola Maragno\cr Politecnico di Milano\cr Maintainer: Paola Maragno\cr E-Mail: <paola.maragno@mail.polimi.it>
}
